Separated Components into individual folders
